+++ aiccs-20191015/alibabacloud_aiccs20191015/models/_cancel_ai_call_details_request.py
@@ -11,6 +11,7 @@ class CancelAiCallDetailsRequest(DaraModel):
         self,
         batch_id: str = None,
         detail_id_list: List[str] = None,
+        encryption_type: int = None,
         owner_id: int = None,
         phone_numbers: List[str] = None,
         resource_owner_account: str = None,
@@ -19,6 +20,7 @@ class CancelAiCallDetailsRequest(DaraModel):
     ):
         self.batch_id = batch_id
         self.detail_id_list = detail_id_list
+        self.encryption_type = encryption_type
         self.owner_id = owner_id
         self.phone_numbers = phone_numbers
         self.resource_owner_account = resource_owner_account
@@ -39,6 +41,9 @@ class CancelAiCallDetailsRequest(DaraModel):
 
         if self.detail_id_list is not None:
             result['DetailIdList'] = self.detail_id_list
+
+        if self.encryption_type is not None:
+            result['EncryptionType'] = self.encryption_type
 
         if self.owner_id is not None:
             result['OwnerId'] = self.owner_id
@@ -64,6 +69,9 @@ class CancelAiCallDetailsRequest(DaraModel):
 
         if m.get('DetailIdList') is not None:
             self.detail_id_list = m.get('DetailIdList')
+
+        if m.get('EncryptionType') is not None:
+            self.encryption_type = m.get('EncryptionType')
 
         if m.get('OwnerId') is not None:
             self.owner_id = m.get('OwnerId')