+++ aiccs-20191015/alibabacloud_aiccs20191015/models/_import_one_task_phone_number_request.py
@@ -9,6 +9,7 @@ from darabonba.model import DaraModel
 class ImportOneTaskPhoneNumberRequest(DaraModel):
     def __init__(
         self,
+        encryption_type: int = None,
         out_id: str = None,
         owner_id: int = None,
         phone_number: str = None,
@@ -17,6 +18,7 @@ class ImportOneTaskPhoneNumberRequest(DaraModel):
         task_id: int = None,
         variables: Dict[str, Any] = None,
     ):
+        self.encryption_type = encryption_type
         self.out_id = out_id
         self.owner_id = owner_id
         # This parameter is required.
@@ -35,6 +37,9 @@ class ImportOneTaskPhoneNumberRequest(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.encryption_type is not None:
+            result['EncryptionType'] = self.encryption_type
+
         if self.out_id is not None:
             result['OutId'] = self.out_id
 
@@ -60,6 +65,9 @@ class ImportOneTaskPhoneNumberRequest(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('EncryptionType') is not None:
+            self.encryption_type = m.get('EncryptionType')
+
         if m.get('OutId') is not None:
             self.out_id = m.get('OutId')