+++ aiccs-20191015/alibabacloud_aiccs20191015/models/_query_ai_call_detail_page_request.py
@@ -13,6 +13,7 @@ class QueryAiCallDetailPageRequest(DaraModel):
         call_result: str = None,
         called_number: str = None,
         detail_ids: List[int] = None,
+        encryption_type: int = None,
         end_calling_time: int = None,
         end_imported_time: int = None,
         major_intent: str = None,
@@ -33,6 +34,7 @@ class QueryAiCallDetailPageRequest(DaraModel):
         self.call_result = call_result
         self.called_number = called_number
         self.detail_ids = detail_ids
+        self.encryption_type = encryption_type
         self.end_calling_time = end_calling_time
         self.end_imported_time = end_imported_time
         self.major_intent = major_intent
@@ -70,6 +72,9 @@ class QueryAiCallDetailPageRequest(DaraModel):
 
         if self.detail_ids is not None:
             result['DetailIds'] = self.detail_ids
+
+        if self.encryption_type is not None:
+            result['EncryptionType'] = self.encryption_type
 
         if self.end_calling_time is not None:
             result['EndCallingTime'] = self.end_calling_time
@@ -131,6 +136,9 @@ class QueryAiCallDetailPageRequest(DaraModel):
 
         if m.get('DetailIds') is not None:
             self.detail_ids = m.get('DetailIds')
+
+        if m.get('EncryptionType') is not None:
+            self.encryption_type = m.get('EncryptionType')
 
         if m.get('EndCallingTime') is not None:
             self.end_calling_time = m.get('EndCallingTime')